student_managment_sys/requirements.txt
fastapi
uvicorn[standard]
SQLmodel
sqlalchemy
sqlite-utils
passlib[bcrypt]
python-jose[cryptography]
python-multipart
